Add workflow to set tag based off MODULE.bazel (#30)
This adds in a new workflow which will tag a repo based on the
MODULE.bazel version. This will trigger when things are merged with
main. If the tag is already set then it will fail. It gets called by
other repos using a workflow call. Planning to implement on guardrail
and pattern_tools first. Testing worked on guardrail.

---------

Co-authored-by: ThomasGira <thomas.gira@patternlabs.tech>